backend/auth/requirements.txt
boto3==1.26.137
PyJWT==2.8.0